requirements.txt
pymongo
python-dotenv
google-cloud-speech
google-auth          
pyaudio
google-genai
pyserial
sounddevice
torch
transformers
pillow